Circumvent issues with flask_restx Api object storing the app
Changes the Api object to use a blueprint instead of waiting for the
app to be given. This has two known benefits IMO. One, you ensure the
Api object isn't deeply associated with a particular app instance,
because the Api code has a design flaw where they store the app
instance. Two, you can init the api in the exact location it's
created.

The intended goal of this change is towards a less global testing
setup.